[WIP] Resolved circular dependency
