+++ great_expectations/datasource/data_connector/inferred_asset_s3_data_connector.py
@@ -4,9 +4,7 @@ from typing import List, Optional
 import great_expectations.exceptions as ge_exceptions
 from great_expectations.core.batch import BatchDefinition
 from great_expectations.core.batch_spec import PathBatchSpec, S3BatchSpec
-from great_expectations.datasource.data_connector.file_path_data_connector import (
-    FilePathDataConnector,
-)
+from great_expectations.datasource.data_connector import ConfiguredAssetS3DataConnector
 
 try:
     import boto3
@@ -79,7 +77,7 @@ class InferredAssetS3DataConnector(InferredAssetFilePathDataConnector):
         )
 
         self._bucket = bucket
-        self._prefix = self.sanitize_prefix_for_s3(prefix)
+        self._prefix = ConfiguredAssetS3DataConnector.sanitize_prefix_for_s3(prefix)
         self._delimiter = delimiter
         self._max_keys = max_keys
 
@@ -92,22 +90,6 @@ class InferredAssetS3DataConnector(InferredAssetFilePathDataConnector):
             raise ImportError(
                 "Unable to load boto3 (it is required for InferredAssetS3DataConnector)."
             )
-
-    @staticmethod
-    def sanitize_prefix_for_s3(text: str) -> str:
-        """
-        Takes in a given user-prefix and cleans it to work with file-system traversal methods
-        (i.e. add '/' to the end of a string meant to represent a directory)
-
-        Customized for S3 paths, ignoring the path separator used by the host OS
-        """
-        path_parts = text.split("/")
-        if not path_parts:  # Empty prefix
-            return text
-        elif "." in path_parts[-1]:  # File, not folder
-            return text
-        else:  # Folder, should have trailing /
-            return text.rstrip("/") + "/"
 
     def build_batch_spec(self, batch_definition: BatchDefinition) -> S3BatchSpec:
         """
